+++ edr_plugin/__init__.py
@@ -7,7 +7,7 @@ from qgis.PyQt.QtWidgets import QAction
 
 from edr_plugin.gui import EdrDialog
 from edr_plugin.gui.browser_panel_models import SavedQueriesItemProvider
-from edr_plugin.gui.coveragejson_drop_handler import CoverageJSONDropHandler, CoverageJSONItemProvider
+from edr_plugin.gui.coveragejson_file_loaders import CoverageJSONDropHandler, CoverageJSONItemProvider
 from edr_plugin.utils import icon_filepath
 from edr_plugin.utils.communication import UICommunication
 from edr_plugin.visualization import EdrLayerManager
@@ -38,7 +38,7 @@ class EDRPlugin:
         self.communication = UICommunication(self.iface, self.PLUGIN_NAME)
         self.actions = []
         self.coveragejson_drop_handler = CoverageJSONDropHandler(self.layer_manager)
-        self.item_provider = CoverageJSONItemProvider(self.layer_manager)
+        self.coveragejson_browser_item_provider = CoverageJSONItemProvider(self.layer_manager)
 
     def add_action(
         self,
@@ -79,7 +79,7 @@ class EDRPlugin:
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
         self.add_action(self.plugin_icon_path, text=self.PLUGIN_NAME, callback=self.run, parent=self.iface.mainWindow())
         self.iface.registerCustomDropHandler(self.coveragejson_drop_handler)
-        QgsApplication.dataItemProviderRegistry().addProvider(self.item_provider)
+        QgsApplication.dataItemProviderRegistry().addProvider(self.coveragejson_browser_item_provider)
 
     def unload(self):
         """Removes the plugin menu item and icon from QGIS GUI."""
@@ -89,7 +89,7 @@ class EDRPlugin:
         # remove the toolbar
         del self.toolbar
         self.iface.unregisterCustomDropHandler(self.coveragejson_drop_handler)
-        QgsApplication.dataItemProviderRegistry().removeProvider(self.item_provider)
+        QgsApplication.dataItemProviderRegistry().removeProvider(self.coveragejson_browser_item_provider)
 
     def ensure_main_dialog_initialized(self):
         if self.main_dialog is None:
