+++ edr_plugin/gui/coveragejson_file_loaders.py
